Call MHD_run only from mainloop, one thread
This fixes some crashes where libmicrohttpd freed the connection
object while it was still in the epoll set. The crash happened
when the epoll event was processed on the freed connection object.

With the MHD_USE_EPOLL_INTERNAL_THREAD set, libmicrohttpd runs in
its own thread. Running one thread with to nested epoll calls and
nonblocking file descriptors seams simpler.